refactor: #13: rename IDatabase file
