+++ rapiddb/src/traits/i_database.rs
