+++ rapiddb/src/traits/mod.rs
@@ -1,7 +1,7 @@
 //! RapidDB Traits
 
-mod idatabase;
+mod i_database;
 mod i_async_database;
 
-pub use idatabase::IDatabase;
+pub use i_database::IDatabase;
 pub use i_async_database::IAsyncDatabase;
